+++ florian-geron/app/root.tsx
@@ -32,7 +32,7 @@ export function Layout({ children }: { children: React.ReactNode }) {
   const navLinks = [
     { to: "/", label: "About Me" },
     { to: "/timeline", label: "Timeline" },
-    // { to: "/contact", label: "Contact" },
+    { to: "/contact", label: "Contact" },
   ];
 
   return (
